+++ src/falconpy/ml_exclusions.py
@@ -1,0 +1,125 @@
+"""
+ _______                        __ _______ __        __ __
+|   _   .----.-----.--.--.--.--|  |   _   |  |_.----|__|  |--.-----.
+|.  1___|   _|  _  |  |  |  |  _  |   1___|   _|   _|  |    <|  -__|
+|.  |___|__| |_____|________|_____|____   |____|__| |__|__|__|_____|
+|:  1   |                         |:  1   |
+|::.. . |   CROWDSTRIKE FALCON    |::.. . |    FalconPy
+`-------'                         `-------'
+
+OAuth2 API - Customer SDK
+
+ml_exclusions - Falcon Machine Learning Exclusions API Interface Class
+
+This is free and unencumbered software released into the public domain.
+
+Anyone is free to copy, modify, publish, use, compile, sell, or
+distribute this software, either in source code form or as a compiled
+binary, for any purpose, commercial or non-commercial, and by any
+means.
+
+In jurisdictions that recognize copyright laws, the author or authors
+of this software dedicate any and all copyright interest in the
+software to the public domain. We make this dedication for the benefit
+of the public at large and to the detriment of our heirs and
+successors. We intend this dedication to be an overt act of
+relinquishment in perpetuity of all present and future rights to this
+software under copyright law.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
+IN NO EVENT SHALL THE AUTHORS BE LIABLE FOR ANY CLAIM, DAMAGES OR
+OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE,
+ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
+OTHER DEALINGS IN THE SOFTWARE.
+
+For more information, please refer to <https://unlicense.org>
+"""
+# pylint: disable=C0103
+from ._util import service_request, parse_id_list, force_default
+from ._service_class import ServiceClass
+
+
+class ML_Exclusions(ServiceClass):   # pylint: disable=C0103
+    """The only requirement to instantiate an instance of this class
+       is a valid token provided by the Falcon API SDK OAuth2 class, a
+       existing instance of the authentication class as an object or a
+       valid set of credentials.
+    """
+    def getMLExclusionsV1(self: object, ids) -> dict:
+        """Get a set of ML Exclusions by specifying their IDs"""
+        # [GET] https://assets.falcon.crowdstrike.com/support/api/swagger.html#/ml-exclusions/getMLExclusionsV1
+        id_list = str(parse_id_list(ids)).replace(",", "&ids=")
+        target_url = f"{self.base_url}/policy/entities/ml-exclusions/v1?ids={id_list}"
+        header_payload = self.headers
+        returned = service_request(caller=self,
+                                   method="GET",
+                                   endpoint=target_url,
+                                   headers=header_payload,
+                                   verify=self.ssl_verify
+                                   )
+        return returned
+
+    def createMLExclusionsV1(self: object, body: dict) -> dict:
+        """Create the ML exclusions"""
+        # [POST] https://assets.falcon.crowdstrike.com/support/api/swagger.html#/custom-ioa/create-rule-groupMixin0
+        target_url = f"{self.base_url}/policy/entities/ml-exclusions/v1"
+        header_payload = self.headers
+        body_payload = body
+        returned = service_request(caller=self,
+                                   method="POST",
+                                   endpoint=target_url,
+                                   body=body_payload,
+                                   headers=header_payload,
+                                   verify=self.ssl_verify
+                                   )
+        return returned
+
+    @force_default(defaults=["parameters"], default_types=["dict"])
+    def deleteMLExclusionsV1(self: object, ids, parameters: dict = None) -> dict:
+        """Delete the ML Exclusions by ID."""
+        # [DELETE] https://assets.falcon.crowdstrike.com/support/api/swagger.html#/ml-exclusions/deleteMLExclusionsV1
+        id_list = str(parse_id_list(ids)).replace(",", "&ids=")
+        target_url = f"{self.base_url}/policy/entities/ml-exclusions/v1?ids={id_list}"
+        header_payload = self.headers
+        parameter_payload = parameters
+        returned = service_request(caller=self,
+                                   method="DELETE",
+                                   endpoint=target_url,
+                                   params=parameter_payload,
+                                   headers=header_payload,
+                                   verify=self.ssl_verify
+                                   )
+        return returned
+
+    def updateMLExclusionsV1(self: object, body: dict) -> dict:
+        """Update the ML Exclusions"""
+        # [PATCH] https://assets.falcon.crowdstrike.com/support/api/swagger.html#/ml-exclusions/updateMLExclusionsV1
+        target_url = f"{self.base_url}/policy/entities/ml-exclusions/v1"
+        header_payload = self.headers
+        body_payload = body
+        returned = service_request(caller=self,
+                                   method="PATCH",
+                                   endpoint=target_url,
+                                   body=body_payload,
+                                   headers=header_payload,
+                                   verify=self.ssl_verify
+                                   )
+        return returned
+
+    @force_default(defaults=["parameters"], default_types=["dict"])
+    def queryMLExclusionsV1(self: object, parameters: dict = None) -> dict:
+        """Search for ML Exclusions."""
+        # [GET] https://assets.falcon.crowdstrike.com/support/api/swagger.html#/ml-exclusions/queryMLExclusionsV1
+        target_url = f"{self.base_url}/policy/queries/ml-exclusions/v1"
+        header_payload = self.headers
+        parameter_payload = parameters
+        returned = service_request(caller=self,
+                                   method="GET",
+                                   endpoint=target_url,
+                                   headers=header_payload,
+                                   params=parameter_payload,
+                                   verify=self.ssl_verify
+                                   )
+        return returned